Remove FutureTable and create a new TokenTable
